Widen g_MpSetup.options to u64; fold portoptions in; 64-bit playlist
The MP-options bitmask was a full 32-bit field (MPOPTION_GOLDENEYE claimed
bit 31). Widen g_MpSetup.options to u64 unconditionally, reclaiming bits
32-63 for port-only options.

- options is u64 in memory and on the wire. SVC_STAGE_START /
  SVC_LOBBY_STATE / CLC_ADMIN_SETUP serialize it via netbufWrite/ReadU64;
  netlobbystate.options widened to match. NET_PROTOCOL_VER 35 -> 37.
- Fold the portoptions overflow word into options bits 32-63.
  MPOPTION_NODOORS moves to bit 32 (0x0000000100000000ULL); the
  portoptions struct field and its separate wire/save sites are removed.
- mpsetups.bin: options saved as a 64-bit inline word replacing the old
  32-bit options + 32-bit portoptions tail (net-zero on the 80-byte
  block). v5 saves migrate the old portoptions tail into bits 32-63 on
  load. MPSETUP_VERSION 5 -> 6.
- Menu: high-word checkboxes use menuhandlerMpCheckboxPortOption,
  retargeted to options bits 32-63 (param3 << 32), since menuitem.param3
  is only 32-bit.
- Dedicated-server playlist / admin path is 64-bit aware: playlistentry
  mp_options/mask are u64; options use a dedicated 64-bit namedoption
  table (NODOORS added); playlistApply is authoritative for high-word
  options (no sticky stopgap).
- Fix /status and admin /show prints (%08x -> %016llx). Docs updated.

Wire (v37) and save (v6) formats changed; not yet built/verified.

+++ port/include/net/playlist.h
@@ -25,8 +25,8 @@ struct playlistentry {
 	s16 stagenum;            // STAGE_MP_* or PLAYLIST_RANDOM_STAGE
 	s8  scenario;            // MPSCENARIO_* or PLAYLIST_RANDOM_SCENARIO
 	s8  weaponpreset;        // g_MpWeaponPresets index, -2 = built-in by name lookup
-	u32 mp_options;          // override bits
-	u32 mp_options_mask;     // which mp_options bits are authoritative
+	u64 mp_options;          // override bits (64-bit; high bits 32-63 are port-only options like MPOPTION_NODOORS)
+	u64 mp_options_mask;     // which mp_options bits are authoritative
 	u8  scorelimit;
 	u8  timelimit;
 	u16 teamscorelimit;
@@ -96,13 +96,15 @@ void playlistDumpToChat(void);
 
 // Name -> id lookups over the playlist parser's tables, for the admin `set`
 // command. Stage/scenario/bot-difficulty return -1 if the name is unknown;
-// option returns the MPOPTION_* bit (0 if unknown). playlistAllOptionBits
-// returns the OR of every settable option bit.
+// option returns the MPOPTION_* bit (0 if unknown — every real option bit is
+// nonzero). The option bit is 64-bit so high-word options (e.g. MPOPTION_NODOORS
+// at bit 32) are representable. playlistAllOptionBits returns the OR of every
+// settable option bit.
 s32 playlistLookupStage(const char *name);
 s32 playlistLookupScenario(const char *name);
 s32 playlistLookupBotDiff(const char *name);
-u32 playlistLookupOption(const char *name);
-u32 playlistAllOptionBits(void);
+u64 playlistLookupOption(const char *name);
+u64 playlistAllOptionBits(void);
 
 // Append one entry to the playlist file the server last loaded (for the admin
 // `saverotation` command). Returns 0 on success, -1 if no file is known or it
